Typesccriptize AppSettingsPageContents
